split into smaller components and added tests
