+++ eventkit_cloud/ui/static/ui/app/index.js
@@ -9,7 +9,7 @@ import LoginPage from './components/auth/LoginPage'
 import Loading from './components/auth/Loading'
 import Logout from './containers/logoutContainer'
 import About from './components/About'
-import DataPackPage from './components/DataPackPage'
+import DataPackPage from './components/DataPackPage/DataPackPage'
 import Export from './components/Export'
 import CreateExport from './components/CreateExport'
 import ExportAOI from './components/ExportAOI'
